refactor: access repository
