--- server/src/interfaces/audit.interface.ts
@@ -1,14 +1,0 @@
-import { DatabaseAction, EntityType } from 'src/enum';
-
-export const IAuditRepository = 'IAuditRepository';
-
-export interface AuditSearch {
-  action?: DatabaseAction;
-  entityType?: EntityType;
-  userIds: string[];
-}
-
-export interface IAuditRepository {
-  getAfter(since: Date, options: AuditSearch): Promise<string[]>;
-  removeBefore(before: Date): Promise<void>;
-}
