+++ server/src/repositories/index.ts
@@ -1,7 +1,6 @@
 import { IAlbumUserRepository } from 'src/interfaces/album-user.interface';
 import { IAlbumRepository } from 'src/interfaces/album.interface';
 import { IAssetRepository } from 'src/interfaces/asset.interface';
-import { IAuditRepository } from 'src/interfaces/audit.interface';
 import { IConfigRepository } from 'src/interfaces/config.interface';
 import { ICronRepository } from 'src/interfaces/cron.interface';
 import { ICryptoRepository } from 'src/interfaces/crypto.interface';
@@ -78,6 +77,7 @@ export const repositories = [
   //
   AccessRepository,
   ActivityRepository,
+  AuditRepository,
   ApiKeyRepository,
 ];
 
@@ -85,7 +85,6 @@ export const providers = [
   { provide: IAlbumRepository, useClass: AlbumRepository },
   { provide: IAlbumUserRepository, useClass: AlbumUserRepository },
   { provide: IAssetRepository, useClass: AssetRepository },
-  { provide: IAuditRepository, useClass: AuditRepository },
   { provide: IConfigRepository, useClass: ConfigRepository },
   { provide: ICronRepository, useClass: CronRepository },
   { provide: ICryptoRepository, useClass: CryptoRepository },
