+++ server/src/types.ts
@@ -3,6 +3,7 @@ import { Permission } from 'src/enum';
 import { AccessRepository } from 'src/repositories/access.repository';
 import { ActivityRepository } from 'src/repositories/activity.repository';
 import { ApiKeyRepository } from 'src/repositories/api-key.repository';
+import { AuditRepository } from 'src/repositories/audit.repository';
 
 export type AuthApiKey = {
   id: string;
@@ -16,6 +17,7 @@ export type RepositoryInterface<T extends object> = Pick<T, keyof T>;
 export type IActivityRepository = RepositoryInterface<ActivityRepository>;
 export type IAccessRepository = { [K in keyof AccessRepository]: RepositoryInterface<AccessRepository[K]> };
 export type IApiKeyRepository = RepositoryInterface<ApiKeyRepository>;
+export type IAuditRepository = RepositoryInterface<AuditRepository>;
 
 export type ActivityItem =
   | Awaited<ReturnType<IActivityRepository['create']>>
