+++ server/test/repositories/audit.repository.mock.ts
@@ -1,4 +1,4 @@
-import { IAuditRepository } from 'src/interfaces/audit.interface';
+import { IAuditRepository } from 'src/types';
 import { Mocked, vitest } from 'vitest';
 
 export const newAuditRepositoryMock = (): Mocked<IAuditRepository> => {
